Security config added
